backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "langchain[ollama]>=1.2.13",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
]
